+++ app/src/main/java/studio/nodroid/bloodpressurehelper/model/User.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class User(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val name: String,
-    val weight: Int = 0
+    val weight: Double = 0.0
 )
